+++ spec/support/test_definition.rb
@@ -1,7 +1,0 @@
-module TestDefinition
-  extend Taskinator::Definition
-
-  def do_task(*args)
-  end
-
-end